ipywidgets: working on writing a widgets manager -- this commit mainly involves some actions refactoring
